Implemented Preliminary Word Detection
Modified hallu_detect.py to identify which parts of the transcript are hallucinated if word_detect is set to True.

This piggy backs off the existing detection mechanism, by using the segments that triggered the detector.

hallu_detect_utils.py contains the functions that implement this.

transcribe.py has also been updated with the new flag.